Assertions and helper functions
Added assertions to the Quadtree_remove test

Added documentation to the update_square helper function in SerialSkipQuadtree.c

The update_square function isn't in Quadtree.h like the other helper functions because it is somewhat time-consuming and expensive, unlike the other helper functions, which are short computations. This means that, depending on the specific implementation, update_square may or may not change -- or even be needed.